LayoutTests/inspector/model/dom-node-expected.txt

Tests for the WI.DOMNode model object.


== Running test suite: WI.DOMNode
-- Running test case: WI.DOMNode.attributes
id="test-id"
class="test-class"
data-item="test-data"
